refactor(deep-research): report generator
