+++ packages/multimodal/deep-research-agent/src/index.ts
@@ -14,3 +14,4 @@ export * from './tools/deep-dive-tool';
 // Export utilities
 export * from './utils/content-processor';
 export * from './utils/search-optimizer';
+export * from './report/report-generator';
